+++ common/components/inputs/DateSelection/RangeButton.tsx
@@ -1,23 +1,18 @@
 import classNames from 'classnames';
 import React from 'react';
-import { lineColorBackground, lineColorDarkBorder } from '../../../styles/general';
+import { lineColorDarkBorder } from '../../../styles/general';
 import { useDelimitatedRoute } from '../../../utils/router';
 import { buttonHighlightConfig } from '../styles/inputStyle';
 
-interface ButtonProps
+interface RangeButtonProps
   extends React.DetailedHTMLProps<
     React.ButtonHTMLAttributes<HTMLButtonElement>,
     HTMLButtonElement
   > {
   children: string | React.ReactNode;
-  isFullWidth?: boolean;
 }
 
-export const DatePickerButton: React.FC<ButtonProps> = ({
-  children,
-  isFullWidth = false,
-  ...props
-}) => {
+export const RangeButton: React.FC<RangeButtonProps> = ({ children, ...props }) => {
   const { line } = useDelimitatedRoute();
 
   return (
@@ -26,8 +21,7 @@ export const DatePickerButton: React.FC<ButtonProps> = ({
       className={classNames(
         'inline-flex  items-center self-stretch border-l bg-black bg-opacity-10 px-3 py-1 text-sm font-medium text-white text-opacity-90 shadow-sm  hover:bg-opacity-5 focus:bg-opacity-0 focus:outline-none ',
         line && buttonHighlightConfig[line],
-        lineColorDarkBorder[line ?? 'DEFAULT'],
-        isFullWidth ? 'w-auto' : 'w-fit'
+        lineColorDarkBorder[line ?? 'DEFAULT']
       )}
       {...props}
     >
